Commit working code before I lose it to git. I am paranoid
