+++ cmd/dgraph/group.conf
@@ -1,2 +1,2 @@
-// Default formula for getting group.
-default: attribute % 10 + 3
+// Default formula for getting group where fp is the fingerprint of a predicate.
+default: fp % 10 + 3
