+++ cmd/dgraph/main.go
@@ -68,7 +68,9 @@ var (
 	cpuprofile  = flag.String("cpu", "", "write cpu profile to file")
 	memprofile  = flag.String("mem", "", "write memory profile to file")
 	schemaFile  = flag.String("schema", "", "Path to schema file")
-	closeCh     = make(chan struct{})
+	groupConf   = flag.String("groups", "groups.conf",
+		"Path to config file with group <-> predicate mapping.")
+	closeCh = make(chan struct{})
 
 	groupId uint64 = 0 // ALL
 )
@@ -614,7 +616,7 @@ func main() {
 
 	posting.InitIndex(ps)
 	posting.Init()
-	worker.ParseGroupConfig("group.conf")
+	worker.ParseGroupConfig(*groupConf)
 
 	var ws *worker.State
 	if groupId != 0 { // HACK: This will currently not run.
