+++ worker/conf_test.go
@@ -3,14 +3,23 @@ package worker
 import "testing"
 
 func TestGroups(t *testing.T) {
-	ParseGroupConfig("group_test1.conf")
+	ParseGroupConfig("group_tests/defaultmissing.conf")
 	gid := group("film.actor.film")
 	if gid != 7 {
 		t.Errorf("Expected groupId to be: %v. Got: %v", 7, gid)
 	}
 
 	groupConfig = config{}
-	ParseGroupConfig("group_test.conf")
+	ParseGroupConfig("group_tests/defaultwrongseq.conf")
+
+	groupConfig = config{}
+	ParseGroupConfig("group_tests/wrongsequence.conf")
+
+	groupConfig = config{}
+	ParseGroupConfig("group_tests/defaultright.conf")
+
+	groupConfig = config{}
+	ParseGroupConfig("group_tests/rightsequence.conf")
 	gid = group("type.object.name.en")
 	if gid != 1 {
 		t.Errorf("Expected groupId to be: %v. Got: %v", 1, gid)
