--- worker/group_test1.conf
@@ -1,2 +1,0 @@
-// Default formula for getting group.
-default: attribute % 10 + 3
