+++ worker/group_tests/defaultmissing.conf
@@ -3,5 +3,4 @@
 1: type.object.name.en
 2: type.object.name*, film.performance.*
 
-// Default formula for getting group.
-default: attribute % 10 + 3
+// Default config missing, should error out.
