+++ worker/group_tests/defaultright.conf
@@ -1,0 +1,3 @@
+// Default formula for getting group.
+// Everything ok should pass.
+default: fp % 10
