+++ worker/group_tests/defaultwrongseq.conf
@@ -1,0 +1,8 @@
+// If * is specified prefix matching would be done.
+0: _uid_
+1: type.object.name.en
+2: type.object.name*, film.performance.*
+
+// Default formula for getting group where fp is the fingerprint of a predicate.
+// 4 > 2 the last groupId should error out.
+default: fp % 10 + 4
