+++ worker/group_tests/wrongsequence.conf
@@ -1,0 +1,9 @@
+// If * is specified prefix matching would be done.
+0: _uid_
+1: type.object.name.en
+
+// Group id not sequential should error out.
+7: type.object.name*, film.performance.*
+
+// Default formula for getting group where fp is the fingerprint of a predicate.
+default: fp % 10 + 4
